Semi-werkend Auto pick systeem
