+++ Racegame versie 0.1/RaceGameExample/PickerP1.resx
@@ -117,26 +117,4 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <assembly alias="System.Windows.Forms" name="System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" />
-  <data name="Racebaan" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Resources\Racebaan.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="Racebaan1" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Resources\Racebaan1.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="Splash" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Resources\Splash.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="BMW M6 GT3" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Resources\BMW M6 GT3.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="Ferari 458 Italia" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Resources\Ferari 458 Italia.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="Mercedes AMG GTR" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Resources\Mercedes AMG GTR.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="Porsche 911 GTR RS" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Resources\Porsche 911 GTR RS.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
 </root>
