+++ Racegame versie 0.1/RaceGameExample/Racegame versie 0.1.csproj
@@ -47,11 +47,23 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Car.cs" />
-    <Compile Include="Form1.cs">
+    <Compile Include="Game.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Form1.Designer.cs">
-      <DependentUpon>Form1.cs</DependentUpon>
+    <Compile Include="Game.Designer.cs">
+      <DependentUpon>Game.cs</DependentUpon>
+    </Compile>
+    <Compile Include="PickerP1.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="PickerP1.Designer.cs">
+      <DependentUpon>PickerP1.cs</DependentUpon>
+    </Compile>
+    <Compile Include="PickerP2.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="PickerP2.Designer.cs">
+      <DependentUpon>PickerP2.cs</DependentUpon>
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
@@ -61,8 +73,14 @@
     <Compile Include="Splash Screen.Designer.cs">
       <DependentUpon>Splash Screen.cs</DependentUpon>
     </Compile>
-    <EmbeddedResource Include="Form1.resx">
-      <DependentUpon>Form1.cs</DependentUpon>
+    <EmbeddedResource Include="Game.resx">
+      <DependentUpon>Game.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="PickerP1.resx">
+      <DependentUpon>PickerP1.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="PickerP2.resx">
+      <DependentUpon>PickerP2.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
@@ -91,6 +109,14 @@
     <None Include="Resources\Racebaan.png" />
   </ItemGroup>
   <ItemGroup>
+    <Content Include="BMW-M6-GT3.png" />
+    <Content Include="Ferarri458Italia.png" />
+    <Content Include="MercedesAMG-GTR.png" />
+    <Content Include="Porsche911-GT3-RS.png" />
+    <Content Include="Resources\BMW M6 GT3.png" />
+    <Content Include="Resources\Ferari 458 Italia.png" />
+    <Content Include="Resources\Mercedes AMG GTR.png" />
+    <Content Include="Resources\Porsche 911 GTR RS.png" />
     <Content Include="Resources\Racebaan1.png" />
     <Content Include="Resources\Splash.png" />
   </ItemGroup>
